added login feature

+++ base/views.py
@@ -1,5 +1,11 @@
 from django.shortcuts import render, redirect
-from .models import Room
+from django.contrib.auth.models import User
+from django.http import HttpResponse
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
+from django.db.models import Q
+from .models import Room, Topic
 from .forms import RoomForm
 
 # rooms=[]
@@ -10,9 +16,49 @@ from .forms import RoomForm
 #     {'id': 3, 'name': "Let's learn Javascript"}, 
 # ]
 
+def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User doesn't exist")
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "Username or Password does not match")
+
+    context={}
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
 def home(request):
-    rooms = Room.objects.all()
-    context = {'rooms': rooms}
+    
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
+
+    rooms = Room.objects.filter(
+        Q(topic__name__icontains=q) |
+        Q(name__icontains=q) |
+        Q(description__icontains=q)
+        )
+    topics = Topic.objects.all()
+    room_count = rooms.count()
+
+    context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
     return render(request, 'base/home.html', context)
 
 def room(request, pk):
@@ -26,6 +72,8 @@ def room(request, pk):
 
     return render(request, 'base/room.html', context)
 
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
 
@@ -38,9 +86,15 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    
+    if request.user != room.host:
+        return HttpResponse("You're not allowed to edit this room")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -51,8 +105,14 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You're not allowed to delete this room")
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
